voice_agent/requirements.txt
# Core dependencies (matching XRAY_Agent versions)
fastapi==0.104.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart>=0.0.6
uvicorn==0.24.0

# LiveKit Agents with Google Gemini Live API - Updated to latest version
livekit-agents[google, deepgram]==1.0.17

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# n8n integration (no external package needed - we use HTTP)
